move NotificationWatcher to component
