+++ src/app/admin/Layout.vue
@@ -1,37 +1,26 @@
 <template>
+<div>
   <router-view></router-view>
+  <notification-watcher />
+</div>
 </template>
 
 <script>
 import { db, firestore } from "~/plugins/firebase.js";
+import NotificationWatcher from "./NotificationWatcher";
 
 export default {
+  components: {
+    NotificationWatcher,
+  },
   data() {
-    return {
-      message_detacher: () => {},
-      uid: this.$store.getters.uidAdmin,
-      notifications: [],
-      watchingMessage: false,
-    }
+    return {};
   },
   async created() {
-    this.watchingMessage = false;
-    this.message_detacher = db
-      .doc(`admins/${this.uid}/private/notification`)
-      .onSnapshot(notification => {
-        if (notification.exists) {
-          const notification_data = notification.data();
-          this.notifications.push(notification_data);
-          if ((this.$route.path.indexOf(notification_data.path) > -1) &&
-              notification_data.sound && this.watchingMessage){
-            this.soundPlay();
-          }
-        }
-        this.watchingMessage = true;
-      });
+    console.log("created");
   },
   destroyed() {
-    this.message_detacher();
+    console.log("destroyed");
   },
 }
